+++ packages/expo-clipboard/android/src/main/java/expo/modules/clipboard/ClipboardModule.kt
@@ -1,0 +1,43 @@
+package expo.modules.clipboard
+
+import android.content.Context
+import android.content.ClipData
+import android.content.ClipboardManager
+
+import org.unimodules.core.ExportedModule
+import org.unimodules.core.ModuleRegistry
+import org.unimodules.core.Promise
+import org.unimodules.core.interfaces.ExpoMethod
+
+class ClipboardModule(context: Context) : ExportedModule(context) {
+
+  override fun getName(): String {
+    return NAME
+  }
+
+  override fun onCreate(moduleRegistry: ModuleRegistry) {
+    ClipboardEventEmitter(context, moduleRegistry)
+  }
+
+  @ExpoMethod
+  fun getStringAsync(promise: Promise) {
+    val clipboard = context.getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
+    val clip = clipboard.getPrimaryClip()
+    if (clip != null && clip.getItemCount() >= 1) {
+        promise.resolve(clip.getItemAt(0).getText());
+    } else {
+        promise.resolve("");
+    }
+  }
+
+  @ExpoMethod
+  fun setString(promise: Promise, content: String) {
+    val clipboard = context.getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
+    val clip = ClipData.newPlainText(null, content)
+    clipboard.setPrimaryClip(clip)
+  }
+
+  companion object {
+    private val NAME = "ExpoClipboard"
+  }
+}